Test SRIOV port move operations with PCI conflicts
This patch tests cold migration, unshelve and evacuate in a situation
where the existing port binding's pci_slot would cause a conflict on
the destination compute node. While cold migration and evacuation
work correctly, in the unshelve case the pci_slot is not updated,
resulting in two instances attempting to consume the same PCI device.
This "passed" in the functional tests, but with a real libvirt this
would obviously explode.

Related: bug 1851545

Change-Id: Ib81532dc1e6dd85822e38eb1785ffb7162d2a84d