Remove unused DeleteFromSelect db api method
Noticed this while I was working on archive_deleted_rows when the code
comment didn't appear to be true and was confusing.

Change I4f2e1632aac96afa7ce06908dc30c0716bcb87b2 removed the last usage
of the DeleteFromSelect method, so remove the unused code to avoid
future confusion.

Change-Id: I4c7574898b64c8f8172ddc84ced86b8a70bf39ff