vmware: OptionValue.value cannot be Boolean
With changing the SOAP library backing oslo.vmware [1], boolean types
will be represented as such in the generated XML while the previous SOAP
library always generated strings. For OptionValue objects, the vCenter
complains about an InvalidArgument when we send a boolean instead of a
string for the value attribute. Therefore, we change all occurrences to
strings.

[1] https://specs.openstack.org/openstack/oslo-specs/specs/victoria/oslo-vmware-soap-library-switch.html

Change-Id: I7659a48f0eeb28cb396e8476a3e2c4b7f027c0bb