[func test] move port creation to the NeutronFixture
The port resource request tests needs some extra ports and networks that
so far was created in the test class setUp(). For the pps support we need
to be able to create NeutronFixture subclass and that to work properly
first we need to encapsulate the port creation into the fixture. We also
took the opportunity to move the resulting new NeutronFixture closer to
the tests that are depending on it.

Change-Id: I85dd8a6af1486de2935b6d1c55cdfe4abf75c19b
blueprint: qos-minimum-guaranteed-packet-rate