Fix bond_mode enum 802.1ad -> 802.3ad
This seems to me a clerical error made a long time ago in the spec [1].
The 802.1ad values does not seems to be a valid bonding mode but 802.3ad
does.

This patch fixes the schema in the nova doc. No test are changed as nova
does not generate this part of the network metadata.

[1] https://specs.openstack.org/openstack/nova-specs/specs/kilo/approved/metadata-service-network-info.html#rest-api-impact

Change-Id: I0055d13b055e34372a8186008ba75be68aa2edf9
Closes-Bug: #1925388