Stop leaking ceph df cmd in RBD utils
If the ceph df command fails in the get_pool_info
method of RBD utils the actual command executed
if seen by the users in the fault error message.

This hides the command behind a StorageError
exception and logs the exception instead of leaking
it to the users.

Change-Id: I6e3a73f2e04d1a7636daf96d5af73c9cf2fbe220
Closes-Bug: 1926978
(cherry picked from commit 86af7feed06f08ddb3ef65122089216708d53a06)