image_meta: Provide image_ref as the id when fetching from instance
At present various users of instance.image_meta expect it to be a fully
populated ImageMeta object. However when the instance is pulled from the
database the ImageMeta object is only partially populated via
instance.system_metadata with a limited set of attributes such as
min_ram, min_disk, disk_format and container_format. This can result in
NotImplementedErrors being raised by o.vo as seen in bug #1928063 when
an attempt is made to access common attributes such as id or name.

To avoid this the id is now provided by instance.image_ref when
available.

The majority of the change is test fallout and ensuring that valid UUIDs
are provided as instance.image_ref and used by instance.image_meta.

Related-Bug: #1928063
Change-Id: I55d66c3a6cbd50da90065f4a58f77b5cd29ce9ea