libvirt: Set driver_iommu when attaching virtio devices to SEV instance
As called out in the original spec [1] virtio devices attached to a SEV
enabled instance must have the iommu attribute enabled. This was done
within the original implementation of the spec for all virtio devices
defined when initially spawning the instance but does not include volume
and interfaces that are later hot plugged.

This change corrects this for both volumes and nics and in doing so
slightly refactors the original designer code to make it usable in both
cases.

[1] https://specs.openstack.org/openstack/nova-specs/specs/train/implemented/amd-sev-libvirt-support.html#proposed-change

Closes-Bug: #1930734
Change-Id: I11131a3f90b8af85e7151b519fb26d225629c391
(cherry picked from commit 4d8bf15fec15dc3416023e577e0f2c277c216506)