zuul: Skip block migration with attached volumes tests due to bug #1931702
Bug #1931702 details soft lockups reported within the guest OS during
live migration with block migration and a volume attached. These lockups
then causing the request to detach the volume as part of the cleanup to
fail. For the time being we should skip these tests until the underlying
issue is resolved.

Related-Bug: #1931702
Change-Id: I7c1a647fb840fce98672a8429d554dd399cd13b7