Allow X-OpenStack-Nova-API-Version header in CORS
By default, we don't currently allow the Nova microversion
header for CORS requests.  It should be something that is
included out of the box because it's part of the core API.

Deployers can workaround this by overriding allow_headers,
but we should provide a better experience out of the box.

Closes-Bug: #1931908
Change-Id: Idf4650f36952331f02d7512580c21451f3ee3b63