Add func test for nova-manage db archive_deleted_rows --before
This is inspired by database query filtering that occurs when --before
is passed to nova-manage db archive_deleted_rows. There is a comparison
WHERE deleted_at < before and not all database records populate it.

This verifies that no records are missed due to that comparision when
archiving with --before.

Related-Bug: #1751192

Change-Id: I44b79cfb236e94444740e32b0dd0a2344c29f340