api: Add an update method for server-groups
This new API endpoint allows to add and/or remove members from a
server-group. We found this to be necessary, because instances might get
created without a server-group, but later need an HA-partner and
re-installing would mean downtime or too much effort.

The endpoint checks the validity of the policy is still given after all
changes are done and strives for idempotency in that it allows
removal/addition of already removed/added instance uuids to accommodate
for requests built in parallel.

TODO:
	* api-request docs

Change-Id: I30d5d1dc3a41553b4336aad3877018989159495c
(cherry picked from commit 7220be3968ee1dd257c9add88228cc5bb9857795)