api: Add an update method for server-groups
This new API endpoint allows to add and/or remove members from a
server-group. We found this to be necessary, because instances might get
created without a server-group, but later need an HA-partner and
re-installing would mean downtime or too much effort.

The endpoint checks the validity of the policy is still given after all
changes are done and strives for idempotency in that it allows
removal/addition of already removed/added instance uuids to accommodate
for requests built in parallel.

We do not allow a server to join a second server-group if it already
joined one. The whole change to the server-group is errored out if any
of the servers are already in another group to make it easier in
validation and either apply the whole change requested by the user or
nothing.

When we add/remove a server to/from a server-group, we have to update
the server's RequestSpec's `instance_group` attribute, because this is
used during scheduling when resizing a server to record a list of
appropriate hosts for the instance.

Only members/admins can change the server-groups.

TODO:
	* api-request docs

Change-Id: I30d5d1dc3a41553b4336aad3877018989159495c
Incorporated-Change-Id: Ibc78cbd80ae49df9be60f6d9eda56a47d9ffe39e
Incorporated-Change-Id: I6a06e4b4a8c22737c20ab51e85ceb2bf98082b26
Incorporated-Change-Id: Ic193dd3c59bc717ba5329f63054297f44127d76d
Co-Authored-By: Jakob Karge <jakob.karge@sap.com>