objects: Fix VIFMigrateData.supports_os_vif_delegation setter
We're using a 'profile' getter/setter in the 'VIFMigrateData' object to
ensure we transform the JSON encoded string we store in the database to
an actual dictionary on load and vice versa on save. However, because
the getter is returning a new object (constructed from 'json.loads')
rather than a reference to the original data (which is a string),
modifying this object doesn't actually change the underlying data in the
object. We were relying on this broken behavior to set the
'supports_os_vif_delegation' attribute of the 'VIFMigrateData' object
and trigger the delegated port creation introduced in change
I11fb5d3ada7f27b39c183157ea73c8b72b4e672e, which means that code isn't
actually doing anything yet. Resolve this.

Change-Id: I362deb1088c88cdcd8219922da9dc9a01b10a940
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Related-Bug: #1734320
Related-Bug: #1815989