api: Align availability zone info with forced host
Users can create a server like so:

  $ openstack server create --availability-zone az:host ...

This is a historical way to request that an instance be scheduled to a
specific host and it causes the scheduler to be bypassed. However, no
validation of this availability zone-host combo takes place. The host
could in fact belong to a different availability zone. If it does, we'll
end up in a very odd situation whereby the RequestSpec record for the
instance will record the availability zone requested by the user at
create time, but the Instance record itself will record the availability
zone of the host on which the instance was scheduled. This leads to even
more confusing behavior when we attempt to do something like live
migrate the instance since the RequestSpec record, with its original and
possibly invalid availability zone information, is used. The
'AvailabilityZoneFilter' will fail an error message like the following:

  Availability Zone 'foo' requested. ... has AZs: bar

but the 'openstack server list --long' command will show a non-foo value
for the availability zone column.

The solution is simple: when given an availability zone-host combo, make
sure the availability zone requested matches that of the host (or, more
specifically, the host is a member of the host aggregates that form the
availability zone [1]). If not, simply ignore the requested availability
zone information in favour of using the availability zone of the host,
logging a warning just for record keeping purposes. This is deemed
preferable to failing with HTTP 400 (Bad Request) since what users are
really requesting by using this was to schedule to a specific host: the
availability zone portion of the request is really irrelevant and just
an artifact of this legacy mechanism to request hosts. If users wish to
truly validate a host-availability zone combo, they can use the 'host'
field introduced in microversion 2.74 along with the 'availability_zone'
field:

  $ openstack server create --availability-zone az --host host ...

Conflicts:
  nova/compute/api.py

NOTE(stephenfin): Conflicts are trivial and due to the absence of change
I81fec10535034f3a81d46713a6eda813f90561cf ("Remove references to
'instance_type'") which we don't want to backport here.

[1] https://docs.openstack.org/nova/latest/admin/aggregates.html

Change-Id: Iac0e634e66cd4e150a50935cf635f626fc11b70e
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Closes-Bug: #1934770
(cherry picked from commit 8f21ee42bd66b62e75e14acf4e91b786d78b9168)