vmware: Do not reraise exception on DRS override removal
We're currently unable to remove a DRS override as our SOAP library is
unable to create an appropriate request accepted by the vCenter.
Therefore, we still allow the resize to work for now and have to
manually remove the override later on.

We keep the code in, so we get a reminder in the logs and sentry, that
this needs fixing. We just don't fail the resize for it.

Change-Id: I4d344347860c7d97d6f4b2e68d9bbac069d71b74
(cherry picked from commit d7b157fa56f6cf55812552ab9a07c9a745d17a4d)