[func test] refactor interface attach with qos
To make the existing negative detach test more generic the logic is
changed from increasing the resource request of the port over the
resource inventory to decrease the resource inventory below the port
request. This is needed as the way to increasing the port request will
be different when the extended resource request extension is enabled.

Change-Id: Iec9d3bb4fef7b6b18a13517872f46afcce02520c
blueprint: qos-minimum-guaranteed-packet-rate