Parse extended resource request from the port data
The format of the value of the resource_request field of the port has
been changed by the port-resource-request-groups Neutron API extension.
This patch adds a new factory method for RequestGroup to parse the data
and create the RequestGroup objects corresponding to the request. Also
the neutron interface of nova changed to use the new factory if the new
neutron API extension is present.

Note that we have to keep the old logic in place to support the scenario
when Nova is upgraded first and therefore Neutron is still using the old
resource_request format.

Also note that some of the so far skipped functional tests started to
pass as they are negative tests resulting in an unsuccessful operation due to
not related reasons and the code already works good enough to hit those
negative code paths.

blueprint: qos-minimum-guaranteed-packet-rate
Change-Id: If7b80c8725d9a8183d2df05c824461e8ee5f45d0